Studiolight: Free Resources for Unused Lights.
Studio lights based on image resoureces are kept, even when only
displayed as an icon. When having many studio lights configured
this can lead to allocating a lot of memory that are actually not used.

This patch free image resources when only icons are requested.
Studio lights that are used in a viewport the image resources are kept.

**TODO**

- [ ] freeing happens from multiple threads so should the studio lights
    needs to be guarded
- [ ] Same mechanism can be done for GPU textures as they don't need
    the images after the gpu textures are constructed.