Fix #116183: Mirror sometimes chooses the wrong vertex
The non-topology option of `EDBM_verts_mirror_cache_begin_ex` only uses
the vertices position to determine whether it is a mirror. However,
some vertices may have the same position. In these cases, another value
can be used to compare.

This commit proposes using the normal and connected edges of a vertex
to determine the best vertex in a group of vertices with the same
position.