Grid-to-points node for converting a grid to a point cloud.
This dumps most available grid information into a point cloud, creating
a point for each voxel and/or tile, with anonymous attributes for the
voxel value and various other properties.

This node is mostly useful for debugging and visualizing grid data,
at least until built-in viewer nodes and spreadsheet support are
available.