Fix #105548: I18n: many keymap sections cannot be translated
In the keymap user preferences, shortcuts are grouped according to a
hierarchy of topics. Some of them cannot be translated, as they do not
get extracted by the i18n script. The issue happens in
`walk_keymap_hierarchy()`. If the item in the hierarchy is not a
string, it is skipped. This happens for tools defining their own
keymap, because they use a function instead of a string as the name.

In turn, the issue is in `keymap_hierarchy.generate()`. If run in the
interface, the hierarchy will contain only strings as topic names.
However, if run in the background, some of them will still be
functions, because they are not initialized.

This commit makes keyconfig initialization run in the background as
well.