Tests: Add bpy_prop_collection foreach_getset tests
Tests the `foreach_get` and `foreach_set` methods of
`bpy_prop_collection`.

The built-in Python sequences of `List` and `Tuple` are tested as well
as sequences which implement the buffer protocol.

When testing buffers, it is not knowable which buffer should be
considered compatible with a property, so the tests only check that at
least one was considered compatible.

Buffer formats with standard size little-endian and standard size
big-endian prefixes are also tested, though neither of these are
expected to be compatible with the current `foreach_getset`
implementation which does not support prefix characters.

The tests are separated into subclasses that aim to each test a
different case of access to the property in
`rna_access.cc#rna_raw_access`:
1. Properties likely to have raw array access (mesh attributes)
2. Properties whose item property pointer can be retrieved, but do not
have raw array access
3. Properties whose item property pointer cannot be retrieved

There is an additional test subclass that specifically tests the runtime
collection property subclass `bpy_prop_collection_idprop`.