Clearer API for updating the nodes modifier after tree changes.
The `MOD_nodes_update_interface` updates ID properties as well as
updating bakes and tagging for depsgraph recalc. When menu enum items
change the properties needed to be updated for correct menus, but not
recalculate the modifier. Since enum changes are not covered by the
`interface_changed` flag, the update was running more often, with a
flag to indicate if the modifier should be recalculated.

To make the API clearer, the `MOD_nodes_update_interface` has been split
into two functions:
- MOD_nodes_update_properties: rebuilds the ID properties
- MOD_nodes_update_properties_and_recalc: rebuild ID properties and then
    tag for modifier evaluation.