WIP Attempt to deduplicate RNA enums for WM events.
Not fully functional as-is.

Likely not what we want to do in the end, due to it adding too much
verbosity to the UI (in keymap context, the general category,
mouse, ndof, etc. is already known, so having it listed again in all and
every event label is... annoying).

Related to PR !116390.