Sculpt: refactor gesture operations to allow cancelling.
The sculpt gesture project operation fulfills the criteria for
checking shape key locks, which implies being able to cancel
the operation if the check fails.

Implementing this requires splitting the gesture begin callback
into two parts that are called before and after the undo push
begin call, and supporting a return value to allow the callback
to signal that the operation should be cancelled.