Vulkan: Fix: Layout Transition
When an image is used for storage, its layout is always "general" outside of the render pass.
When using ShaderRead-Only, transitions can cause system instability,
so leave it in General-layout for read-only access (no validation errors).
InputAttachment embeds transitions within the render pass, but requires binding.
However, if the image is outside of the render pass, it can have any layout.
In this way, it can be said that writing with Subpass is advantageous over writing with Storage
not only in terms of pixel access efficiency but also in terms of layout transition.