Vulkan: Fix: Layout Transition
When running ClearOp, its initial layout must be Attachments.
This is not the basic handling of layout transitions.
Therefore, record the barrier just before starting the render pass.
Running a barrier within a render pass increases dependencies for that pass.
This is a trade-off. In this case, it is necessary to clear it immediately after the render pass starts,
so I think it is better to make the transition outside the render pass.