Fix #112599: Blender freezes for half a minute when opening File Browser
Considering the disadvantage of waiting for
`Shortcut->Resolve(0, SLR_NO_UI | SLR_UPDATE)` to find the path that
may have been moved or renamed, it seems better to just not call this
function.

A disadvantage of this change is that, if the target directory is not
found (because it was moved or renamed), Blender's File Browser will
not access the shortcut nor display it if it is filtered.

One advantage is that the user will not have to mysteriously wait
minutes for the inaccessible directory of a shortcut to be found.