Fix compilation error of GPU tests with per-suit executables
Add the common GPU test sources to every individual test
executable. Additionally, do not create executable for the
common sources.

For the optimal compile time the common part might need to
be moved to its own target, similar to the bf_testing_main,
but leaving that for the further work.