Fix #116807: Default value crash for large multidimensional bpy property definitions
The functions for creating bpy vector properties used fixed size
`default_value` arrays on the stack. The array was not large enough to
store all the default values for larger multidimensional properties,
which would write past the end of the array and start corrupting the
stack, leading to crashes.

The maximum total length of a bpy vector property is 32768 (32*32*32)
which is a little large of an array to be putting on the stack, so the
`default_value` array is now allocated when default values need to be
parsed.

To simplify allocating arrays of the correct size,
#bpy_prop_array_length_parse now fills in `array_len_info->len_total`
when parsing a sequence.