Fix: off-by-1 in curves_selection.cc#contains
Consider an IndexRange of 8 elements starting at 0:
`range.start()` will be `0` and `range.last()` will be `7`.

The size of the range is not `range.last() - range.start()` because that
will be 7, but rather `range.one_after_last() - range.start()`.

The code was using `last()` when it should have been using
`one_after_last()`, which would cause the very last element of each
2048-element section (as per the grain size) of `varray` to be skipped.