GPv3: Crash deleting end points
Crash is observed when `dst_curve` count is same as original curve.
`type_count`in runtime struct of `dst_curve` are not updated after copying
curve attributes hence the crash/assert. In other case (when source and
dst curve_nums are different), type_count is updated when dst.finish()
is called in `gather_attributes()`