I18n: make the FPS menu in Format Panel react to language change
The Frame Rate menu uses a custom label to display the selected frame
rate. The function uses a caching mechanism to avoid recalculating on
each draw if the value has not changed, but that means that it does
not update if the translation language is changed.

This commit removes this mechanism so that the translation is correct
even after a language change.