Vulkan: Fix Anisotropic Filtering
When anisotropic filtering is enabled on a sampler its value must be
between 1 and 16. In blender it is possible to set a value lower than
1. This would trigger a validation error in Vulkan.

This PR updates this by clamping the user pref setting into the valid
range. It seems like the previous implementation was always incorrect
as it would set an invalid value or 1.0.