Fix #114692: Sculpt on certain modifiers does not update normals
Caused bu 5af8b839cf.

This solution tags both mesh positions and normals for update when
drawing happens from the modifier stack evaluation result.

This is needed because the sculpt session modifies original mesh
positions, and the modifier stack is not guaranteed to modify them
further (and hence tag normals for update).