Refactor: Store a 'Mesh' in the editmesh snap cache and use it for snapping
Create a temporary Mesh representing the edited mesh and use it in the
edit mesh snapping system.

By using a Mesh object for snapping, we remove a considerable amount of
code and use a more optimized version of snapping.

This simplifies the code and makes it easier to implement new features.

The only downside is that more memory is required once a temporary mesh
is created.