WIP: Fix possible evaluation of hidden objects
Under certain conditions it is possible that some objects which are
hidden and which visibility is not evaluated are still evaluated when
their dependencies are modified.

One of the contributing aspects to the issue was that detection of
visible IDs might not be fully correct when some of the IDs in the
dependency chain are ignored early on and their IDNode::is_enabled_on_eval
is never evaluated and left to default true, forcing other objects
in the scene to be evaluated.

This change ensures that the is_enabled_on_eval is calculated at least
once during the dependency graph construction.