Fix possible evaluation of hidden objects
Under certain conditions it is possible that some objects which are
hidden and which visibility is not evaluated are still evaluated when
their dependencies are modified.

Disallow relation from object to the collection it belongs to to force
object evaluation, and fully rely on the base flag exclusion bits to
optimize evaluation of hidden and restricted objects. This is only done
for collections which are reachable from layer collections. This way
collections for which visiiblity an not be controlled from the user
interface are evaluated the same exact way.