Vulkan: Refactor: Deferred Rendering
A naive fix regarding closure_count. When doing Deferred-Rendering,
closure_count will be the number of `subpass` if a fallback is required.
Therefore, you need to pass that value when setting the render pass.
This fix should allow us to run `subpass` up to 8 times.
Whether this is a correct modification for the API should be considered separately.