Render: Skip writing output if no Composite node exists
This patch changes the render pipeline such that it doesn't write the
render result if the render pipeline is a compositor one whose node tree
does not have an output Composite node.

This is done to avoid double outputs when users use the File Output node
for saving the results instead. Additionally, it fixes crashes and empty
outputs when the node tree does not reference any render layers and has
no Composite output.

The render pipeline early fails if neither a File Output or a Composite
output node exists, so an output is guaranteed to always exist.

Fixes #115983 and #57699.