IK: distinguish 'off' from '0% influence' in IK constraints
Depending on the situation, the IK solver will use the influence slider
in different ways. If there is no pole vector, it will interpolate between
the target position and the bone's current position. This means that an
influence of 0% will still fully run the IK solver, which overwrites the
bone's transform matrix.

This commit introduces a distinction between '0% influence' and 'muted'.
The former still behaves as above, and the latter now actually skips the
IK solver for that bone chain.

There is a difference in target positions between '0% influence' and 'muted'
when it comes to tree IK (i.e. multiple bones with IK constraints, where the
IK chains at some point overlap at a common parent bone).

This should fix #115915.