IK: draw relationship line when influence is 0%, hide when OFF
When IK constraints are enabled was recently changed. This commit updates
the overlay drawing code to match.

The relationship line, between the start and the end of the IK chain, is
now drawn whenever the IK system is active for that chain. This means that
it's drawn even at 0% influence, but not when the constraint is OFF/muted.