Fix: Memory leak report with static select engine data
Now that the select engine data uses non-trivial objects in its global
data, storing the data at the global scope causes more trouble than it's
worth. Instead use the construct on first use idiom to make it clear
when the data is first constructed or destructed. Though this struct
probably shouldn't be static at all, it does simplify memory management
as well, removing the need to manually clear the arrays causing the issue.