Placeholders for non-pointer properties in `uiItemPointerR`.
Showing pointer properties will automatically show a placeholder string
if the property is not set. The `uiItemPointerR` function can also be
used with string properties (e.g. vertex groups, layer names), but in
that case the UI will not automatically show a placeholder string.

This is because the ui_but_placeholder_get uses the property itself as
the source of the pointer type. For string properties there is no struct
type associated with the property and thus no placeholder is generated.
The button function does not know that the property is associated with
a collection.

To make this work the higher-level `uiItemPointerR_prop` function can
determine a placeholder based on the search property instead.